Corrección de errores y controles agregados.

+++ src/main/java/com/remiNorte/app/controllers/RemisController.java
@@ -13,9 +13,11 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.SessionAttributes;
 import org.springframework.web.bind.support.SessionStatus;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import com.remiNorte.app.models.entity.Remis;
 import com.remiNorte.app.models.service.IRemisService;
+import com.remiNorte.app.models.service.IViajeService;
 
 @Controller
 @SessionAttributes("remis")
@@ -23,6 +25,9 @@ public class RemisController {
 
 	@Autowired 
 	private IRemisService remisService;
+	
+	@Autowired
+	private IViajeService viajeService;
 	
 	@RequestMapping(value="/listarRemises", method=RequestMethod.GET)
 	public String listarRemises(Model model) {
@@ -45,7 +50,7 @@ public class RemisController {
 		return "formRemis";
 	}
 		
-	@RequestMapping(value="/formRemis/{id}")
+	@RequestMapping(value="/formRemis/{id}", method=RequestMethod.GET)
 	public String editar(@PathVariable(value="id") Long id,Map<String, Object> model) {
 		Remis remis = null;
 		if (id > 0) {
@@ -54,33 +59,58 @@ public class RemisController {
 			return "redirect:/listarRemises";
 		}
 		model.put("remis", remis);
-		model.put("titulo", "Modificar Remis");
+		model.put("backPage", "/listarRemises");
+		model.put("titulo", "Modificar Remís");
 		return "formRemis";
 	}
 	
 	@RequestMapping(value="/formRemis", method=RequestMethod.POST)
-	public String guardar(@Valid Remis remis, BindingResult result, Model model, SessionStatus status) { //metodo que procesa el formulario
+	public String guardar(@Valid Remis remis, 
+							BindingResult result, 
+							Model model, 
+							SessionStatus status) { //metodo que procesa el formulario
 		
 		if(result.hasErrors()) {
 			model.addAttribute("titulo", "Nuevo Registro");
 			return "formRemis";
 		}
 		
-		remisService.save(remis);
-		status.setComplete(); //elimina el objeto cliente de la sesión.
+		Remis remisAux = null;
+		remisAux = remisService.findByRemPatente(remis.getRemPatente().trim());
 		
-		//model.addAttribute("titulo", "Informacion del Remis");
-		//model.addAttribute("remis", remis);
+		if (remisAux == null || remisAux.getRemId() == remis.getRemId()) {
+			remisService.save(remis);
+			status.setComplete(); //elimina el objeto cliente de la sesión.			
+		} else {
+			result.rejectValue("remPatente","remPatente","Patente ya registrada en otro remís.");
+		}
+		
+		if (result.hasErrors()) {
+			model.addAttribute("remis", remis);
+			model.addAttribute("titulo", "Modificar Remís");
+			model.addAttribute("backPage", "/listarRemises");
+			return "formRemis";
+		}
 		
 		return "redirect:listarRemises";
 	}
 	
 
 	@RequestMapping(value="/eliminarRemis/{id}")
-	public String eliminar(@PathVariable(value="id") Long id) {
+	public String eliminar(@PathVariable(value="id") Long id, 
+							RedirectAttributes flash) {
+		
 		if (id > 0) {
-			remisService.delete(id);
-		}		
+			Long canRemViajes = (long) 0;
+			canRemViajes = viajeService.canRemVia(id);
+			
+			if (canRemViajes == 0) {
+				remisService.delete(id);
+			} else {
+				flash.addFlashAttribute("error", "El remís no se puede eliminar porque tiene viajes asociados.");
+			}
+			
+		}
 		return "redirect:/listarRemises";
 	}
